feat(members): Add `GET /v1/members/{jid}/email-address`
Closes #314.
