+++ docs/openapi/features/profile.yaml
@@ -41,6 +41,25 @@ paths:
       "204": { description: Success }
       "401": { $ref: "../shared.yaml#/components/responses/Unauthorized" }
       "403": { $ref: "../shared.yaml#/components/responses/Forbidden" }
+  get_member_email_address:
+    tags: [Members]
+    summary: Get member email address
+    description: Get a member’s email address.
+    operationId: get_member_email_address
+    security:
+      - BearerAuth: []
+    parameters:
+      - $ref: "members.yaml#/components/parameters/MemberId"
+    responses:
+      "200":
+        description: Success
+        content:
+          application/json:
+            schema:
+              $ref: "../shared.yaml#/components/schemas/EmailAddress"
+              type: [string, "null"]
+      "401": { $ref: "../shared.yaml#/components/responses/Unauthorized" }
+      "403": { $ref: "../shared.yaml#/components/responses/Forbidden" }
   set_member_nickname:
     tags: [Members]
     summary: Set member nickname
