+++ tests/integration/step-ci/members.yaml
@@ -186,7 +186,30 @@ tests:
             status: /2../
             headers:
               Pagination-Item-Count: 13
-
+  setAndGetEmailAddress:
+    name: Set then get a member’s email address
+    steps:
+      - name: Set email address
+        http:
+          method: PUT
+          url: /v1/members/${{ env.MEMBER_JIDS.0 }}/email-address
+          auth:
+            bearer:
+              token: ${{ captures.token }}
+          json: ${{ env.MEMBER_EMAILS.0 }}
+          check:
+            status: /2../
+      - name: Get email address
+        http:
+          method: GET
+          url: /v1/members/${{ env.MEMBER_JIDS.0 }}/email-address
+          auth:
+            bearer:
+              token: ${{ captures.token }}
+          check:
+            status: 200
+            jsonpath:
+              $.: ${{ env.MEMBER_EMAILS.0 }}
 components:
   steps:
     add_member:
