+++ tests/integration/test.env
@@ -2,6 +2,7 @@ PROSE_LOG__LEVEL=trace
 PROSE_SERVER__LOG_LEVEL=debug
 PROSE_LOG__FORMAT=pretty
 PROSE_LOG__WITH_ANSI=true
+#PROSE_LOG__WITH_SPAN_EVENTS=true
 PROSE_DEBUG_USE_AT_YOUR_OWN_RISK__LOG_CONFIG_AT_STARTUP=false
 PROSE_DEBUG_USE_AT_YOUR_OWN_RISK__DETAILED_ERROR_RESPONSES=true
 PROSE_DEBUG_ONLY__INSECURE_PASSWORD_ON_AUTO_ACCEPT_INVITATION=true
